extract openapiMiddleware() into its own file
